Package code under tiledb.video & minor refactoring
